isolate viz plugins
