+++ front/components/markdown/Visualization.tsx
@@ -1,15 +1,20 @@
+import type { LightWorkspaceType } from "@dust-tt/types";
 import { useContext, useMemo } from "react";
+import { visit } from "unist-util-visit";
 
+import { VisualizationActionIframe } from "@app/components/assistant/conversation/actions/VisualizationActionIframe";
 import { MarkdownContentContext } from "@app/components/assistant/markdown/MarkdownContentContext";
 import type { CustomRenderers } from "@app/components/assistant/markdown/RenderMessageMarkdown";
 
 const VISUALIZATION_MAGIC_LINE = "{/** visualization-complete */}";
 
+type PositionType = { start: { line: number }; end: { line: number } };
+
 export function VisualizationBlock({
   position,
   customRenderer,
 }: {
-  position: { start: { line: number }; end: { line: number } };
+  position: PositionType;
   customRenderer?: CustomRenderers;
 }) {
   const { content } = useContext(MarkdownContentContext);
@@ -35,4 +40,84 @@ export function VisualizationBlock({
     complete = true;
   }
   return visualizationRenderer(code, complete, position.start.line);
+}
+
+export function getVisualizationPlugin(
+  owner: LightWorkspaceType,
+  agentConfigurationId: string,
+  conversationId: string,
+  messageId: string
+) {
+  const VisualizationPlugin = ({ position }: { position: PositionType }) => {
+    return (
+      <VisualizationBlock
+        position={position}
+        customRenderer={{
+          visualization: (
+            code: string,
+            complete: boolean,
+            lineStart: number
+          ) => {
+            return (
+              <VisualizationActionIframe
+                owner={owner}
+                visualization={{
+                  code,
+                  complete,
+                  identifier: `viz-${messageId}-${lineStart}`,
+                }}
+                key={`viz-${messageId}-${lineStart}`}
+                conversationId={conversationId}
+                agentConfigurationId={agentConfigurationId}
+              />
+            );
+          },
+        }}
+      />
+    );
+  };
+
+  return VisualizationPlugin;
+}
+
+export function visualizationDirective() {
+  return (tree: any) => {
+    visit(tree, ["containerDirective"], (node) => {
+      if (node.name === "visualization") {
+        const data = node.data || (node.data = {});
+        data.hName = "visualization";
+        data.hProperties = {
+          position: node.position,
+        };
+      }
+    });
+  };
+}
+
+export function sanitizeVisualizationContent(str: string) {
+  const lines = str.split("\n");
+
+  let openVisualization = false;
+  for (let i = 0; i < lines.length; i++) {
+    // (2) Replace legacy <visualization> XML tags by the markdown directive syntax for backward
+    // compatibility with older <visualization> tags.
+    if (lines[i].trim() === "<visualization>") {
+      lines[i] = ":::visualization";
+    }
+    if (lines[i].trim() === "</visualization>") {
+      lines[i] = ":::";
+    }
+
+    // (3) Prepend closing visualization markdow directive with a magic word to detect that the
+    // visualization is complete solely based on its content during token streaming.
+    if (lines[i].trim().startsWith(":::visualization")) {
+      openVisualization = true;
+    }
+    if (openVisualization && lines[i].trim() === ":::") {
+      lines.splice(i, 0, VISUALIZATION_MAGIC_LINE);
+      openVisualization = false;
+    }
+  }
+
+  return lines.join("\n");
 }
